remove unused parameters from heartbeat callback, add new params that allow changing the interval, and cleanup types so that all heartbeat intervals are uint64_t.

+++ src/api/cpp/GravityHeartbeat.h
@@ -18,7 +18,7 @@ namespace gravity
 //For passing Data to the HB threads.
 struct HBParams {
 	void* zmq_context;
-	int interval_in_microseconds;
+	uint64_t interval_in_microseconds;
 	//unsigned short port;
 	std::string componentID;
 	std::string endpoint;
